Carve `TestModelWeightGraph` and `TestModelCall` out of the monolith (wala/ML#635).
The 11 weight-graph-walk tests and 28 model/layer-call dataflow tests
move verbatim into their classes; the monolith keeps 553, preserving
the 1,040 total.
